ticket-ingestion-service/requirements.txt
# FastAPI and dependencies
fastapi==0.124.4
uvicorn[standard]==0.38.0
pydantic==2.12.5
pydantic-settings==2.12.0

# Google Cloud
google-cloud-pubsub==2.33.0

# Additional dependencies
python-dotenv==1.2.1
python-multipart==0.0.20
